refactor: extract construction core policy maps
