+++ addons/smart_construction_core/core_extension_policy_maps.py
@@ -1,0 +1,472 @@
+# -*- coding: utf-8 -*-
+from __future__ import annotations
+
+ROLE_SURFACE_OVERRIDES = {
+    "business_config_admin": {
+        "label": "业务配置管理员",
+        "landing_scene_candidates": ["projects.list", "projects.ledger", "projects.intake"],
+        "menu_xmlids": [
+            "smart_construction_core.menu_sc_root",
+            "smart_construction_core.menu_sc_business_config_center",
+        ],
+    },
+    "owner": {
+        "landing_scene_candidates": ["projects.list", "projects.intake"],
+        "menu_xmlids": [
+            "smart_construction_core.menu_sc_project_center",
+            "smart_construction_core.menu_sc_contract_center",
+        ],
+    },
+    "pm": {
+        "landing_scene_candidates": ["portal.dashboard", "projects.ledger", "projects.list", "projects.intake"],
+        "menu_xmlids": [
+            "smart_construction_core.menu_sc_project_center",
+            "smart_construction_core.menu_sc_contract_center",
+            "smart_construction_core.menu_sc_cost_center",
+        ],
+        "menu_blocklist_xmlids": ["smart_construction_core.menu_sc_project_manage"],
+    },
+    "finance": {
+        "landing_scene_candidates": ["finance.payment_requests", "projects.ledger", "projects.list"],
+        "menu_xmlids": [
+            "smart_construction_core.menu_sc_finance_center",
+            "smart_construction_core.menu_sc_settlement_center",
+            "smart_construction_core.menu_payment_request",
+        ],
+    },
+    "executive": {
+        "landing_scene_candidates": ["portal.dashboard", "project.management", "projects.list", "projects.ledger", "projects.intake"],
+        "menu_xmlids": [
+            "smart_construction_core.menu_sc_root",
+            "smart_construction_core.menu_sc_projection_root",
+            "smart_construction_core.menu_sc_project_center",
+        ],
+    },
+}
+
+ROLE_GROUPS_EXPLICIT = {
+    "business_config_admin": {
+        "smart_construction_core.group_sc_cap_business_config_admin",
+    },
+    "executive": {
+        "smart_construction_custom.group_sc_role_executive",
+    },
+    "pm": {
+        "smart_construction_custom.group_sc_role_pm",
+        "smart_construction_custom.group_sc_role_project_manager",
+        "smart_construction_custom.group_sc_role_project_user",
+        "smart_construction_core.group_sc_role_project_manager",
+    },
+    "finance": {
+        "smart_construction_custom.group_sc_role_finance",
+        "smart_construction_custom.group_sc_role_payment_manager",
+        "smart_construction_custom.group_sc_role_payment_user",
+        "smart_construction_custom.group_sc_role_payment_read",
+        "smart_construction_core.group_sc_role_finance_manager",
+        "smart_construction_core.group_sc_role_finance_user",
+    },
+}
+
+ROLE_GROUPS_CAPABILITY_FALLBACK = {
+    "pm": {
+        "smart_construction_core.group_sc_cap_project_manager",
+        "smart_construction_core.group_sc_cap_project_user",
+    },
+    "finance": {
+        "smart_construction_core.group_sc_cap_finance_user",
+        "smart_construction_core.group_sc_cap_finance_manager",
+    },
+}
+
+ROLE_PRECEDENCE = ("business_config_admin", "executive", "pm", "finance")
+
+NAV_MENU_SCENE_MAP = {
+    "smart_construction_core.menu_sc_project_initiation": "projects.intake",
+    "smart_construction_core.menu_sc_project_project": "projects.list",
+    "smart_construction_core.menu_sc_project_management_scene": "project.management",
+    "smart_construction_core.menu_sc_project_cost_code": "config.project_cost_code",
+    "smart_construction_core.menu_sc_root": "projects.list",
+    "smart_construction_core.menu_sc_project_dashboard": "projects.dashboard",
+    "smart_construction_core.menu_sc_history_todo": "workspace.home",
+    "smart_construction_core.menu_sc_operating_metrics_project": "dashboard.company",
+    "smart_construction_core.menu_sc_dashboard_cost_cockpit_fact": "cost.control",
+    "smart_construction_core.menu_sc_dictionary": "data.dictionary",
+    "smart_construction_core.menu_payment_request": "finance.payment_requests",
+}
+
+NAV_ACTION_SCENE_MAP = {
+    "smart_construction_core.action_project_initiation": "projects.intake",
+    "smart_construction_core.action_sc_project_list": "projects.list",
+    "smart_construction_core.action_project_dashboard": "projects.dashboard",
+    "smart_construction_core.action_project_dictionary": "data.dictionary",
+    "smart_construction_core.action_project_cost_code": "config.project_cost_code",
+    "smart_construction_core.action_sc_dashboard_cost_cockpit_fact": "cost.control",
+    "smart_construction_core.action_payment_request": "finance.payment_requests",
+    "smart_construction_core.action_payment_request_my": "finance.payment_requests",
+}
+
+NAV_MODEL_VIEW_SCENE_MAP = {
+    ("project.project", "list"): "projects.list",
+    ("project.project", "form"): "projects.intake",
+    ("payment.request", "list"): "finance.payment_requests",
+    ("payment.request", "form"): "finance.payment_requests",
+}
+
+SERVER_ACTION_WINDOW_MAP = {
+    "smart_construction_core.action_exec_structure_entry": "smart_construction_core.action_exec_structure_wbs",
+}
+
+FILE_ATTACHMENT_ALLOWED_MODEL_EXACT = {
+    "payment.ledger",
+    "payment.request",
+    "payment.request.line",
+    "sc.legacy.fund.confirmation.document",
+    "sc.legacy.invoice.tax.fact",
+    "sc.legacy.payment.residual.fact",
+}
+
+FILE_ATTACHMENT_ALLOWED_MODEL_PREFIXES = ("construction.", "project.", "quota.", "sc.", "tender.")
+
+FILE_ATTACHMENT_ALLOWED_LEGACY_MODEL_PREFIXES = ("sc.legacy.",)
+
+FILE_ATTACHMENT_EXCLUDED_MODEL_PREFIXES = (
+    "sc.legacy.",
+    "sc.ops.",
+    "sc.pack.",
+    "sc.scene.",
+    "sc.subscription.",
+    "sc.usage.",
+    "sc.workbench.",
+    "ui.form.",
+)
+
+FILE_UPLOAD_ALLOWED_MODELS = ["project.project", "project.task", "payment.request"]
+
+FILE_DOWNLOAD_ALLOWED_MODELS = ["project.project", "project.task", "payment.request"]
+
+LEGACY_VISIBLE_BUSINESS_COLUMN_LABELS_BY_MODEL = {
+    "project.material.plan": {
+        "legacy_visible_01": "单据状态",
+        "legacy_visible_02": "单据编号",
+        "legacy_visible_03": "单据日期",
+        "legacy_visible_04": "到货时间",
+        "legacy_visible_05": "采购材料名称",
+        "legacy_visible_06": "规格型号",
+        "legacy_visible_07": "单位",
+        "legacy_visible_08": "数量",
+        "legacy_visible_09": "材料别名(设计/清单)",
+        "legacy_visible_10": "备注",
+        "legacy_visible_11": "附件",
+        "legacy_visible_12": "项目名称",
+        "legacy_visible_13": "录入人",
+        "legacy_visible_14": "录入时间",
+        "source_created_by": "录入人",
+        "source_created_at": "录入时间",
+    },
+    "sc.material.inbound": {
+        "legacy_visible_01": "单据状态",
+        "legacy_visible_02": "入库单号",
+        "legacy_visible_03": "单据日期",
+        "legacy_visible_04": "供应商名称",
+        "legacy_visible_05": "材料名称",
+        "legacy_visible_06": "规格型号",
+        "legacy_visible_07": "数量",
+        "legacy_visible_08": "单价",
+        "legacy_visible_09": "税率",
+        "legacy_visible_10": "含税金额",
+        "legacy_visible_11": "入库总数量",
+        "legacy_visible_12": "付款状态",
+        "legacy_visible_13": "已付款金额",
+        "legacy_visible_14": "未付款金额",
+        "legacy_visible_15": "结算状态",
+        "legacy_visible_16": "已结算金额",
+        "legacy_visible_17": "项目名称",
+        "legacy_visible_18": "备注",
+        "legacy_visible_19": "附件",
+        "legacy_visible_20": "录入人",
+        "legacy_visible_21": "录入时间",
+        "legacy_visible_22": "采购人",
+        "source_created_by": "录入人",
+        "source_created_at": "录入时间",
+    },
+}
+
+API_DATA_WRITE_ALLOWLIST = {
+    "project.project": ["name", "description", "date_start"],
+    "project.task": ["name", "description", "date_deadline", "project_id"],
+    "purchase.order.line": ["name", "order_id"],
+    "res.partner": ["name", "email", "phone", "sc_supplier_type", "sc_supplier_type_ids"],
+}
+
+API_DATA_MUTATION_POLICIES = {
+    "sc.legacy.receipt.income.fact": {
+        "allowed_ops": ["create", "write"],
+        "allowed": False,
+        "reason_code": "READONLY_PROJECTION_MUTATION_DENIED",
+        "message": "历史事实投影为只读数据，不允许通过业务办理接口创建或修改。",
+        "source": "smart_construction_core",
+    },
+}
+
+DRAFT_DELETE_ALLOWED_STATES = ("cancel", "cancelled", "draft")
+
+def _state_unlink_policy(
+    model_name: str,
+    business_label: str,
+    allowed_states=DRAFT_DELETE_ALLOWED_STATES,
+    state_field: str = "state",
+):
+    return {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "policy_kind": "state_limited_business_document",
+        "state_field": state_field,
+        "allowed_states": list(allowed_states),
+        "reason_code": "DRAFT_BUSINESS_DOCUMENT_DELETE_ALLOWED",
+        "message": f"允许删除未形成业务事实的{business_label}；仅限草稿/取消等未提交状态，并继续受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    }
+
+API_DATA_DRAFT_UNLINK_POLICIES = {
+    "construction.contract": _state_unlink_policy("construction.contract", "合同记录"),
+    "construction.contract.income": _state_unlink_policy("construction.contract.income", "收入合同"),
+    "construction.contract.expense": _state_unlink_policy("construction.contract.expense", "支出合同"),
+    "payment.request": _state_unlink_policy("payment.request", "付款申请"),
+    "sc.general.contract": _state_unlink_policy("sc.general.contract", "综合合同"),
+    "sc.expense.claim": _state_unlink_policy("sc.expense.claim", "费用与保证金单据"),
+    "sc.financing.loan": _state_unlink_policy("sc.financing.loan", "融资借款单据"),
+    "sc.invoice.registration": _state_unlink_policy("sc.invoice.registration", "发票登记单据"),
+    "sc.payment.execution": _state_unlink_policy("sc.payment.execution", "付款执行单"),
+    "sc.receipt.income": _state_unlink_policy("sc.receipt.income", "收款收入登记"),
+    "sc.fund.account.operation": _state_unlink_policy("sc.fund.account.operation", "资金账户操作单"),
+    "sc.self.funding.registration": _state_unlink_policy("sc.self.funding.registration", "自筹资金登记"),
+    "sc.tax.deduction.registration": _state_unlink_policy("sc.tax.deduction.registration", "税票抵扣登记"),
+    "sc.settlement.order": _state_unlink_policy("sc.settlement.order", "结算单"),
+    "sc.settlement.adjustment": _state_unlink_policy("sc.settlement.adjustment", "结算调整单"),
+    "project.material.plan": _state_unlink_policy("project.material.plan", "材料计划"),
+    "sc.material.purchase.request": _state_unlink_policy("sc.material.purchase.request", "材料采购申请"),
+    "sc.material.acceptance": _state_unlink_policy("sc.material.acceptance", "材料验收单"),
+    "sc.material.inbound": _state_unlink_policy("sc.material.inbound", "材料入库单"),
+    "sc.material.outbound": _state_unlink_policy("sc.material.outbound", "材料出库单"),
+    "sc.material.rfq": _state_unlink_policy("sc.material.rfq", "材料询比价"),
+    "sc.material.settlement": _state_unlink_policy("sc.material.settlement", "材料结算单"),
+    "sc.material.rental.plan": _state_unlink_policy("sc.material.rental.plan", "材料租赁计划"),
+    "sc.material.rental.order": _state_unlink_policy("sc.material.rental.order", "材料租赁订单"),
+    "sc.material.rental.settlement": _state_unlink_policy("sc.material.rental.settlement", "材料租赁结算"),
+    "sc.labor.plan": _state_unlink_policy("sc.labor.plan", "劳务计划"),
+    "sc.labor.request": _state_unlink_policy("sc.labor.request", "劳务申请"),
+    "sc.labor.usage": _state_unlink_policy("sc.labor.usage", "劳务使用记录"),
+    "sc.labor.settlement": _state_unlink_policy("sc.labor.settlement", "劳务结算"),
+    "sc.labor.price": _state_unlink_policy("sc.labor.price", "劳务价格单"),
+    "sc.equipment.plan": _state_unlink_policy("sc.equipment.plan", "设备计划"),
+    "sc.equipment.request": _state_unlink_policy("sc.equipment.request", "设备申请"),
+    "sc.equipment.usage": _state_unlink_policy("sc.equipment.usage", "设备使用记录"),
+    "sc.equipment.settlement": _state_unlink_policy("sc.equipment.settlement", "设备结算"),
+    "sc.equipment.price": _state_unlink_policy("sc.equipment.price", "设备价格单"),
+    "sc.safety.plan": _state_unlink_policy("sc.safety.plan", "安全方案"),
+    "sc.safety.disclosure": _state_unlink_policy("sc.safety.disclosure", "安全交底"),
+    "sc.safety.issue": _state_unlink_policy("sc.safety.issue", "安全问题"),
+    "sc.safety.patrol.task": _state_unlink_policy("sc.safety.patrol.task", "安全巡检任务"),
+    "sc.quality.issue": _state_unlink_policy("sc.quality.issue", "质量问题"),
+    "sc.quality.rectification": _state_unlink_policy(
+        "sc.quality.rectification",
+        "质量整改记录",
+        ("submitted", "rectifying", "rechecking", "cancel"),
+        state_field="issue_state",
+    ),
+    "sc.quality.recheck": _state_unlink_policy(
+        "sc.quality.recheck",
+        "质量复验记录",
+        ("rectifying", "rechecking", "cancel"),
+        state_field="issue_state",
+    ),
+    "sc.safety.rectification": _state_unlink_policy(
+        "sc.safety.rectification",
+        "安全整改记录",
+        ("submitted", "rectifying", "rechecking", "cancel"),
+        state_field="issue_state",
+    ),
+    "sc.safety.recheck": _state_unlink_policy(
+        "sc.safety.recheck",
+        "安全复验记录",
+        ("rectifying", "rechecking", "cancel"),
+        state_field="issue_state",
+    ),
+    "sc.construction.diary": _state_unlink_policy("sc.construction.diary", "施工日志"),
+    "project.progress.entry": _state_unlink_policy("project.progress.entry", "进度填报"),
+    "project.risk.action": _state_unlink_policy("project.risk.action", "风险措施"),
+    "sc.plan": _state_unlink_policy("sc.plan", "项目计划"),
+    "sc.plan.line": _state_unlink_policy("sc.plan.line", "项目计划明细"),
+    "sc.plan.version": _state_unlink_policy("sc.plan.version", "计划版本"),
+    "sc.plan.report": _state_unlink_policy("sc.plan.report", "计划上报"),
+    "tender.bid": _state_unlink_policy("tender.bid", "投标主单", ("prepare", "estimating")),
+    "tender.doc.purchase": _state_unlink_policy("tender.doc.purchase", "投标文件购买申请"),
+    "tender.doc.review": _state_unlink_policy("tender.doc.review", "投标文件审查"),
+    "tender.guarantee": _state_unlink_policy("tender.guarantee", "投标保证金"),
+}
+
+API_DATA_UNLINK_POLICIES = {
+    "construction.contract": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理合同记录；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "construction.contract.income": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理收入合同记录；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "construction.contract.expense": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理支出合同记录；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "hr.department": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理组织部门；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "payment.request": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理付款申请；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "payment.request.line": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理付款申请明细；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "project.cost.code": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理成本科目；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "project.dictionary": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理业务字典；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "project.task": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许删除任务记录；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "project.tags": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "RELATION_MAINTENANCE_DELETE_ALLOWED",
+        "message": "允许删除项目标签等关系维护数据；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "res.partner": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理客户/供应商资料；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "sc.approval.policy": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理审批策略；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "sc.approval.step": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理审批步骤；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "sc.document.admin.document": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理行政档案；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "sc.hr.payroll.document": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理薪酬档案；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "sc.office.admin.document": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理办公行政资料；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "sc.project.stage.requirement.item": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理阶段要求；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+    "sc.supplier.type": {
+        "allowed": True,
+        "delete_mode": "unlink",
+        "reason_code": "DELETE_POLICY_ALLOWED",
+        "message": "允许业务配置管理员整理供应商类型；仍受模型 ACL 与记录规则约束。",
+        "source": "smart_construction_core",
+    },
+}
+
+API_DATA_UNLINK_POLICIES.update(API_DATA_DRAFT_UNLINK_POLICIES)
+
+API_DATA_UNLINK_ALLOWED_MODELS = list(API_DATA_UNLINK_POLICIES)
+
+MODEL_CODE_MAPPING = {
+    "project": "project.project",
+    "task": "project.task",
+}
+
+CRITICAL_SCENE_TARGET_OVERRIDES = {
+    "projects.list",
+    "projects.detail",
+    "projects.intake",
+    "projects.ledger",
+    "projects.execution",
+    "projects.dashboard",
+    "project.management",
+    "my_work.workspace",
+    "portal.dashboard",
+    "finance.payment_requests",
+}
+
+CRITICAL_SCENE_TARGET_ROUTE_OVERRIDES = {
+    "my_work.workspace": "/my-work",
+}
+
+INDUSTRY_CREATE_FIELD_FALLBACKS = {
+    "project.project": {
+        "selection_defaults": {
+            "privacy_visibility": "followers",
+            "rating_status": "stage",
+            "last_update_status": "to_define",
+            "rating_status_period": "monthly",
+        }
+    }
+}
